Add logic to find the file containing behaviors for parrot
This will first search for a file specified in the environment variable
"PARROT_BEHAVIORS_FILE" if that is not set then a file in the same
directory as the parrot.py file named "parrot_behaviors" will be used.

If the environment variable is set, but the file does not exist parrot
will NOT fall back to using the local file if one exists. This is to
prevent possible confusion. Because parrot cannot provide error messages
on stderr uncaught exceptions will be thrown if the behavior files cannot
be found. This is done to make clear there is an error with the operation
of parrot and not a potentially expected output from a behavior.