Make parrot work!
Tied up all the previous changes into a working tool that can parrot a
command!

A couple bugs were exposed with this.
An existing, but empty behavior file was not properly giving a parrot
  exception and instead was being caught by the json loader.
Python seems to always want to put ./ on the front of a command even
  when the command called is on the path and was not invoked with ./ .
  This caused an issue because the ./ was used as part of determining
  the id of a behavior, which wasn't desirable.
  The solution was to strip ./ iff the command starts with it. This may
  reduce some flexibilty, but it seems unlikely someone would want to
  distinguish between "./<command>" and "<command>" in a behavior file.

An integration test system was setup to test the full run of parrot.
Currently this only has one test as creating the test infrastructure was
a fair bit of work, which is itself a testament to why tools like this
are needed. The integrating tests should serve as both a test of parrot
and and example of how you would use it.